breaks apart AppContext.tsx into AppContext.ts, AppProvider.tsx, types.ts, useAppContext.ts. for react fast refresh to work, components (tsx files) should only export one function, the component. AppContext contains the context. AppProvider is the context provider component. Types are defined in types.ts. useAppContext contains the hook that allows us to use the context in various components.
