+++ vite-react/src/context/AppContext.ts
@@ -1,0 +1,4 @@
+import type { AppState } from "./types";
+import { createContext } from "react";
+
+export const AppContext = createContext<AppState | undefined>(undefined);
