+++ vite-react/src/context/useAppContext.ts
@@ -1,0 +1,10 @@
+import { useContext } from "react";
+import { AppContext } from "./AppContext";
+
+export function useAppContext() {
+  const context = useContext(AppContext);
+  if (!context) {
+    throw new Error("useAppContext must be used within an AppProvider");
+  }
+  return context;
+}
