+++ vite-react/src/views/WelcomeInputView.tsx
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { useAppContext } from "@/context/AppContext";
+import { useAppContext } from "@/context/useAppContext";
 import { useNavigate } from "react-router-dom";
 import { Card, CardContent, CardHeader, CardTitle } from "@/components/ui/card";
 import { Button } from "@/components/ui/button";
